+++ todo/src/App.js
@@ -1,28 +1,21 @@
-import React, { useState } from 'react'
-import { TodoForm } from './components/TodoForm/TodoForm'
-import { TodoList } from './components/TodoList/TodoList'
+import React from 'react'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import { TodoPage } from './pages/TodoPage/TodoPage'
+import { AboutPage } from './pages/AboutPage/AboutPage'
+import { Layout } from './components/Layout/Layout'
 
 export default function App() {
-    const [todoList, setTodoList] = useState([])
-    
-    const addTodo = (name, description) => {
-        if (!name) return
-        const newTodo = {
-            id: Date.now(),
-            name: name,
-            description: description
-        }
-        setTodoList([...todoList, newTodo])
-    }
-    const removeTodo = (item) => {
-        setTodoList(
-            todoList.filter(todo => todo.id !== item.id)
-        )
-    } 
     return (
     <div className='App'>
-        <TodoForm create={addTodo} />
-        <TodoList todos={todoList} deleteTodo={removeTodo}/>
+        <BrowserRouter>
+            <Routes>
+                <Route path='/' element={<Layout />}>
+                    <Route index element={<TodoPage />} />
+                    <Route path='todo' element={<TodoPage />} />
+                    <Route path='about' element={<AboutPage />} />
+                </Route>
+            </Routes>
+        </BrowserRouter>
     </div>
     )
 }
