+++ todo/src/components/Header/Header.module.scss
@@ -5,4 +5,6 @@
   
   height: 80px;
   background-color: $header-color;
+
+  display: flex;
 }
