+++ todo/src/components/Layout/Layout.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Outlet } from 'react-router-dom'
+import { Header } from '../Header/Header'
+
+export const Layout = () => {
+  return (
+    <>
+        <Header />
+        <main>
+            <Outlet />
+        </main>
+    </>
+  )
+}
