+++ todo/src/components/Navbar/Navbar.module.scss
@@ -1,0 +1,11 @@
+.Navbar {
+  position: relative;
+  display: flex;
+  & > ul {
+    list-style: none;
+    display: flex;
+
+    width: 100%;
+    height: 100%;
+  }
+}
