+++ todo/src/pages/AboutPage/AboutPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Layout } from '../../components/Layout/Layout'
+
+export const AboutPage = () => {
+  return (
+    <>
+        <h1>AboutPage</h1>
+    </>
+  )
+}
